requirements.txt
python-docx
pymongo
pprint
numpy
matplotlib